+++ application/src/test/resources/log4j2-test.xml
@@ -15,13 +15,13 @@
     <Logger name="javasabr.mqtt.model.network.impl.DefaultMqttSession" level="DEBUG" additivity="false">
       <AppenderRef ref="BrokerConsoleTest"/>
     </Logger>
-    <Logger name="javasabr.mqtt.service.impl.DefaultIncomingPublishRouter" level="DEBUG" additivity="false">
+    <Logger name="javasabr.mqtt.service.publish.impl.DefaultIncomingPublishRouter" level="DEBUG" additivity="false">
       <AppenderRef ref="BrokerConsoleTest"/>
     </Logger>
-    <Logger name="javasabr.mqtt.service.publish.handler.impl.AbstractIncomingPublishProcessor" level="DEBUG" additivity="false">
+    <Logger name="javasabr.mqtt.service.publish.processor.AbstractIncomingPublishProcessor" level="DEBUG" additivity="false">
       <AppenderRef ref="BrokerConsoleTest"/>
     </Logger>
-    <Logger name="javasabr.mqtt.service.publish.handler.impl.Qos1IncomingPublishProcessor" level="DEBUG" additivity="false">
+    <Logger name="javasabr.mqtt.service.publish.processor.Qos1IncomingPublishProcessor" level="DEBUG" additivity="false">
       <AppenderRef ref="BrokerConsoleTest"/>
     </Logger>
     <Logger name="javasabr.mqtt.service.session.impl.InMemoryMqttSessionService" level="DEBUG" additivity="false">
