+++ core-service/src/main/java/javasabr/mqtt/service/message/handler/impl/PublishMqttInMessageHandler.java
@@ -15,9 +15,7 @@ import javasabr.mqtt.network.impl.ExternalNetworkMqttUser;
 import javasabr.mqtt.network.message.in.PublishMqttInMessage;
 import javasabr.mqtt.network.session.NetworkMqttSession;
 import javasabr.mqtt.service.AuthorizationService;
-import javasabr.mqtt.service.IncomingPublishRouter;
 import javasabr.mqtt.service.MessageOutFactoryService;
-import javasabr.mqtt.service.PublishDataStorage;
 import javasabr.mqtt.service.TopicService;
 import javasabr.mqtt.service.message.validator.PublishMessageExpiryIntervalMqttInMessageFieldValidator;
 import javasabr.mqtt.service.message.validator.PublishPayloadMqttInMessageFieldValidator;
@@ -25,6 +23,8 @@ import javasabr.mqtt.service.message.validator.PublishQosMqttInMessageFieldValid
 import javasabr.mqtt.service.message.validator.PublishResponseTopicMqttInMessageFieldValidator;
 import javasabr.mqtt.service.message.validator.PublishRetainMqttInMessageFieldValidator;
 import javasabr.mqtt.service.message.validator.PublishTopicAliasMqttInMessageFieldValidator;
+import javasabr.mqtt.service.publish.IncomingPublishRouter;
+import javasabr.mqtt.service.publish.PublishDataStorage;
 import javasabr.rlib.common.util.StringUtils;
 import lombok.AccessLevel;
 import lombok.CustomLog;
