--- core-service/src/main/java/javasabr/mqtt/service/publish/handler/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package javasabr.mqtt.service.publish.handler;
-
-import org.jspecify.annotations.NullMarked;
