+++ core-service/src/main/java/javasabr/mqtt/service/publish/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.mqtt.service.publish;
+
+import org.jspecify.annotations.NullMarked;
