+++ core-service/src/main/java/javasabr/mqtt/service/publish/processor/Qos2IncomingPublishProcessor.java
@@ -1,4 +1,4 @@
-package javasabr.mqtt.service.publish.handler.impl;
+package javasabr.mqtt.service.publish.processor;
 
 import javasabr.mqtt.model.MqttUser;
 import javasabr.mqtt.model.QoS;
@@ -18,10 +18,9 @@ import javasabr.mqtt.network.message.in.PublishReleaseMqttInMessage;
 import javasabr.mqtt.network.message.out.MqttOutMessage;
 import javasabr.mqtt.network.session.NetworkMqttSession;
 import javasabr.mqtt.service.MessageOutFactoryService;
-import javasabr.mqtt.service.PublishDispatcher;
-import javasabr.mqtt.service.RetainPublishService;
 import javasabr.mqtt.service.SubscriptionService;
-import javasabr.mqtt.service.publish.handler.PublishHandlingResult;
+import javasabr.mqtt.service.publish.PublishDispatcher;
+import javasabr.mqtt.service.publish.RetainPublishService;
 import lombok.AccessLevel;
 import lombok.CustomLog;
 import lombok.experimental.FieldDefaults;
@@ -113,7 +112,7 @@ public class Qos2IncomingPublishProcessor extends TrackableIncomingPublishProces
       ExternalNetworkMqttUser user,
       NetworkMqttSession session,
       Publish publish,
-      PublishHandlingResult handlingResult) {
+      PublishProcessingResult handlingResult) {
     super.handleError(user, session, publish, handlingResult);
 
     int messageId = publish.messageId();
