+++ core-service/src/main/java/javasabr/mqtt/service/publish/sender/Qos0SubscriberPublishSender.java
@@ -1,4 +1,4 @@
-package javasabr.mqtt.service.publish.handler.impl;
+package javasabr.mqtt.service.publish.sender;
 
 import javasabr.mqtt.model.QoS;
 import javasabr.mqtt.model.publish.OutgoingPublish;
