+++ core-service/src/main/java/javasabr/mqtt/service/publish/sender/TrackableSubscriberPublishSender.java
@@ -1,4 +1,4 @@
-package javasabr.mqtt.service.publish.handler.impl;
+package javasabr.mqtt.service.publish.sender;
 
 import javasabr.mqtt.model.MqttProtocolErrors;
 import javasabr.mqtt.model.MqttUser;
