+++ core-service/src/test/groovy/javasabr/mqtt/service/IntegrationServiceSpecification.groovy
@@ -9,23 +9,23 @@ import javasabr.mqtt.model.MqttVersion
 import javasabr.mqtt.model.QoS
 import javasabr.mqtt.network.MqttConnection
 import javasabr.mqtt.network.handler.NetworkMqttUserReleaseHandler
-import javasabr.mqtt.service.impl.DefaultIncomingPublishRouter
 import javasabr.mqtt.service.impl.DefaultMessageOutFactoryService
-import javasabr.mqtt.service.impl.DefaultPublishDispatcher
 import javasabr.mqtt.service.impl.DefaultTopicService
 import javasabr.mqtt.service.impl.DisabledAuthorizationService
 import javasabr.mqtt.service.impl.InMemoryClientIdRegistry
-import javasabr.mqtt.service.impl.InMemoryRetainPublishService
 import javasabr.mqtt.service.impl.InMemorySubscriptionService
 import javasabr.mqtt.service.message.handler.impl.PublishReleaseMqttInMessageHandler
 import javasabr.mqtt.service.message.out.factory.Mqtt311MessageOutFactory
 import javasabr.mqtt.service.message.out.factory.Mqtt5MessageOutFactory
-import javasabr.mqtt.service.publish.handler.impl.Qos0IncomingPublishProcessor
-import javasabr.mqtt.service.publish.handler.impl.Qos0SubscriberPublishSender
-import javasabr.mqtt.service.publish.handler.impl.Qos1IncomingPublishProcessor
-import javasabr.mqtt.service.publish.handler.impl.Qos1SubscriberPublishSender
-import javasabr.mqtt.service.publish.handler.impl.Qos2IncomingPublishProcessor
-import javasabr.mqtt.service.publish.handler.impl.Qos2SubscriberPublishSender
+import javasabr.mqtt.service.publish.impl.DefaultIncomingPublishRouter
+import javasabr.mqtt.service.publish.impl.DefaultPublishDispatcher
+import javasabr.mqtt.service.publish.impl.InMemoryRetainPublishService
+import javasabr.mqtt.service.publish.processor.Qos0IncomingPublishProcessor
+import javasabr.mqtt.service.publish.processor.Qos1IncomingPublishProcessor
+import javasabr.mqtt.service.publish.processor.Qos2IncomingPublishProcessor
+import javasabr.mqtt.service.publish.sender.Qos0SubscriberPublishSender
+import javasabr.mqtt.service.publish.sender.Qos1SubscriberPublishSender
+import javasabr.mqtt.service.publish.sender.Qos2SubscriberPublishSender
 import javasabr.mqtt.service.session.MqttSessionService
 import javasabr.mqtt.service.session.impl.InMemoryMqttSessionService
 import javasabr.mqtt.test.support.BaseSpecification
